Rename, clean up and refactor Dialog (now MarkerLoader)
